ms compiler

+++ crypto_kem/hqc-128/clean/repetition.c
@@ -26,7 +26,7 @@ void PQCLEAN_HQC128_CLEAN_repetition_code_encode(uint64_t *em, const uint64_t *m
     pos_r = 0;
     for (size_t i = 0; i < VEC_N1_SIZE_64; i++) {
         for (size_t j = 0; j < 64 && pos_r < PARAM_N1N2; j++) {
-            bit = -((m[i] >> j) & 1);
+            bit = (uint64_t) (-(int64_t) ((m[i] >> j) & 1));
             idx_r = (pos_r & 0x3f);
             em[(pos_r >> 6) + 0] ^= (bit & 0x7FFFFFFFUL) << idx_r;
             em[(pos_r >> 6) + 1] ^= (bit & 0x3FFFFFFFUL) >> ((63 - idx_r));
